refactor: Expose what is actually used externally
