+++ src/narwhals/_interchange/__init__.py
@@ -1,4 +1,7 @@
 from __future__ import annotations
 
-from narwhals._interchange.dataframe import unsupported_error as unsupported_error
+from narwhals._interchange.dataframe import (
+    InterchangeFrame as InterchangeFrame,
+    should_interchange as should_interchange,
+)
 from narwhals._interchange.lazyframe import LazyFrame as LazyFrame
