+++ src/narwhals/stable/v1/__init__.py
@@ -7,7 +7,7 @@ import narwhals as nw
 from narwhals import exceptions, functions as nw_f
 from narwhals._exceptions import issue_warning
 from narwhals._expression_parsing import ExprKind, ExprNode, is_expr
-from narwhals._interchange.dataframe import InterchangeFrame, should_interchange
+from narwhals._interchange import InterchangeFrame, should_interchange
 from narwhals._typing_compat import TypeVar, assert_never
 from narwhals._utils import (
     Implementation,
